+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## [1.0.1]
+- Bugfix: fixed link to video
+- Bugfix: fixed redirect to 'edit' after 'create'
+- Bugfix: fixed permission check
+
 ## [1.0.0]
 - Initial Release